requirements.txt
alabaster==0.7.10
arrow==0.10.0
asn1crypto==0.23.0
astroid==1.4.9
Babel==2.5.1
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
BeautifulSoup==3.2.1
cached-property==1.3.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.3.4
cryptography==1.8.1
cssselect==1.0.1
Cython==0.25.2
decorator==4.1.2
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.9
django-braces==1.11.0
django-ckeditor==5.2.1
django-cors-headers==2.0.2
django-debug-toolbar==1.9.1
django-extensions==1.7.7
django-filter==1.0.2
django-formset-js==0.5.0
django-fsm==2.5.0
django-import-export==0.7.0
django-jquery-js==3.1.1
django-localflavor==1.4.1
django-model-utils==2.6.1
django-mptt==0.8.7
django-querysetsequence==0.7.2
django-rest-swagger==2.1.2
django-reversion==2.0.8
django-sslserver==0.19
django-tables2==1.4.2
django-tables2-reports==0.0.10
django-url-filter==0.3.0
django-wkhtmltopdf==3.1.0
django-wysiwyg==0.8.0
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
docutils==0.13.1
enum34==1.1.6
et-xmlfile==1.0.1
Fabric==1.13.1
feedparser==5.2.1
flake8==3.3.0
gunicorn==19.7.1
html5lib==1.0b8
idna==2.6
imagesize==0.7.1
ipaddress==1.0.18
ipdb==0.10.2
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.2.15
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.5
lazy-object-proxy==1.3.1
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mechanize==0.2.5
numpy==1.12.1
odfpy==1.3.6
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.5.0
packaging==16.8
pandas==0.19.2
paramiko==2.4.0
pathlib2==2.3.0
pickleshare==0.7.4
Pillow==4.0.0
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.7.1
py-gfm==0.1.3
pyapi-gitlab==7.8.5
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.5.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.6.5
PyNaCl==1.2.1
pyOpenSSL==17.0.0
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
pyquery==1.2.17
python-dateutil==2.6.1
python-json-logger==0.1.7
python-ldap==2.5.2
python-magic==0.4.13
pytz==2017.3
PyYAML==3.12
qrcode==5.3
reportlab==3.4.0
requests==2.18.4
scandir==1.6
scipy==1.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
sqlparse==0.2.4
suds==0.4
tablib==0.12.1
terminaltables==3.1.0
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.11.15
whitenoise==3.3.1
win-unicode-console==0.5
wkhtmltopdf==0.2
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==0.9.6
xlutils==2.0.0
xlwt==1.2.0
xmltodict==0.10.2
